+++ R/portalcasting.R
@@ -15,19 +15,12 @@
 #' @importFrom utils download.file packageDescription read.csv sessionInfo
 #'  tail unzip write.csv write.table
 #' @importFrom viridis viridis
-#' @importFrom yaml as.yaml yaml.load_file
+#' @importFrom yaml read_yaml write_yaml
 
 #' @title Functions for Portalcasting (Forecasting Portal Rodents)
 #'
-#' @description This package contains the functions used for continuous
-#'  analysis and forecasting of 
-#'  \href{https://portal.weecology.org/}{Portal rodent populations},
-#'  as implemented in the 
-#'  \href{https://bit.ly/2Lo2xnQ}{Portal Predictions repository}. \cr
-#'  Functions in \code{portalcasting} are designed to be portable, allowing
-#'  users to set up a fully-functional replica repository for development
-#'  and testing of new models (a.k.a. \emph{sandboxing}) in the same 
-#'  directory and codebase as the production pipeline. 
+#' @description Functions used for continuous analysis and forecasting of \href{https://portal.weecology.org/}{Portal rodent populations}, as implemented in the \href{https://bit.ly/2Lo2xnQ}{Portal Predictions repository}. \cr \cr
+#'              The code in \code{portalcasting} is designed to be portable, allowing users to set up a fully-functional replica repository for development and testing of new models (a.k.a. \emph{sandboxing}) in the same directory and codebase as the production pipeline. 
 #'
 #' @name portalcasting
 #'
@@ -36,4 +29,3 @@
 #' @keywords package
 #'
 NULL
-